requirements.txt
google-genai>=1.30.0
tqdm>=4.67.0
Pillow>=10.0.0
